feature-BJS2-27252 refactoring
