+++ src/main/java/faang/school/projectservice/mapper/project/ProjectMapper.java
@@ -1,11 +1,11 @@
 package faang.school.projectservice.mapper.project;
 
 
+import faang.school.projectservice.model.dto.ProjectDto;
+import faang.school.projectservice.model.entity.Project;
+import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
-import org.mapstruct.InjectionStrategy;
-import faang.school.projectservice.dto.project.ProjectDto;
-import faang.school.projectservice.model.entity.Project;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE, injectionStrategy = InjectionStrategy.CONSTRUCTOR)
 public interface ProjectMapper {
