+++ src/test/java/faang/school/projectservice/service/project/ProjectServiceTest.java
@@ -1,5 +1,26 @@
 package faang.school.projectservice.service.project;
 
+import faang.school.projectservice.exception.FileTooLargeException;
+import faang.school.projectservice.mapper.project.ProjectMapper;
+import faang.school.projectservice.model.dto.ProjectDto;
+import faang.school.projectservice.model.entity.Project;
+import faang.school.projectservice.repository.ProjectRepository;
+import faang.school.projectservice.service.ProjectService;
+import faang.school.projectservice.service.S3Service;
+import faang.school.projectservice.validator.project.ProjectValidator;
+import jakarta.persistence.EntityNotFoundException;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -7,27 +28,6 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-
-import faang.school.projectservice.service.ProjectService;
-import faang.school.projectservice.service.resource.S3Service;
-import jakarta.persistence.EntityNotFoundException;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.junit.jupiter.MockitoExtension;
-import faang.school.projectservice.dto.project.ProjectDto;
-import faang.school.projectservice.exception.FileTooLargeException;
-import faang.school.projectservice.mapper.project.ProjectMapper;
-import faang.school.projectservice.model.entity.Project;
-import faang.school.projectservice.repository.ProjectRepository;
-import faang.school.projectservice.validator.project.ProjectValidator;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
 
 @ExtendWith(MockitoExtension.class)
 public class ProjectServiceTest {
@@ -66,7 +66,7 @@ public class ProjectServiceTest {
         imageName = "test.png";
         project = new Project();
         project.setId(projectId);
-        projectService.maxFileSize = maxFileSize;
+        projectService.setMaxFileSize(maxFileSize);
         project.setCoverImageId("image-id");
         contentType = "image/png";
         coverImageKey = "unique-key";
